can display groupid

+++ FeiSMS/src/com/njnu/kai/feisms/ContentContactsDetailActivity.java
@@ -21,7 +21,7 @@ public class ContentContactsDetailActivity extends TabActivity implements OnChec
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.content_contacts_detail);
 		
-		int groupId = getIntent().getIntExtra(FeiSMSConst.GROUP_ID, 0);
+		long groupId = getIntent().getLongExtra(FeiSMSConst.GROUP_ID, 0);
 		
 		mIntentSMS = new Intent(this, TabSMSContentActivity.class);
 		mIntentSMS.putExtra(FeiSMSConst.GROUP_ID, groupId);
